# My third attempt at fixing issue 1521 (not being merged due to performance concerns)