--- a/include/simdjson/common_defs.h
+++ b/include/simdjson/common_defs.h
@@ -252,11 +252,19 @@ namespace std {
 #endif
 
 #if SIMDJSON_CPLUSPLUS17
+// if we have C++, then fallthrough is a default attribute
 # define simdjson_fallthrough [[fallthrough]]
-#elif __has_attribute(__fallthrough__)
+// check if we have __attribute__ support
+#elif defined(__has_attribute)
+// check if we have the __fallthrough__ attribute
+#if __has_attribute(__fallthrough__)
+// we are good to go:
 # define simdjson_fallthrough                    __attribute__((__fallthrough__))
-#else
-# define simdjson_fallthrough                    do {} while (0)  /* fallthrough */
+#endif
+#endif
+// on some systems, we simply do not have support for fallthrough, so use a default:
+#ifndef simdjson_fallthrough
+# define simdjson_fallthrough do {} while (0)  /* fallthrough */
 #endif
 
 #endif // SIMDJSON_COMMON_DEFS_H