--- a/include/simdjson/generic/ondemand/json_iterator-inl.h
+++ b/include/simdjson/generic/ondemand/json_iterator-inl.h
@@ -58,17 +58,21 @@ simdjson_warn_unused simdjson_really_inline error_code json_iterator::skip_child
       _depth--;
       if (depth() <= parent_depth) { return SUCCESS; }
       break;
+    case '"':
+      if(*peek() == ':') {
+        // we are at a key!!! This is
+        // only possible if someone searched
+        // for a key and the  key was not found.
+        logger::log_value(*this, "key");
+        advance(); // eat up the ':'
+        break; // important!!!
+      }
     // Anything else must be a scalar value
     default:
-      if(*peek() == ':') {
-        // we are at a key!!!
-        logger::log_value(*this, "key");
-      } else {
-        // For the first scalar, we will have incremented depth already, so we decrement it here.
-        logger::log_value(*this, "skip");
-        _depth--;
-        if (depth() <= parent_depth) { return SUCCESS; }
-      }
+      // For the first scalar, we will have incremented depth already, so we decrement it here.
+      logger::log_value(*this, "skip");
+      _depth--;
+      if (depth() <= parent_depth) { return SUCCESS; }
       break;
   }