The core issue seems to be with skip_child. The skip_child function makes some assumption regarding the state you are in. In particular, it seems that you cannot be right before a key so that the first step of skip_child is to bring you to a key. That's never a problem when the key you are searching for is found. When the key you are searching found is not found, then it is possible that you may end up just before a key.

By convention, keys are at the level of the object, so the code in the main branch and in the 0.9.x is wrong in such cases because skip_child will encounter a key, it will decrement the depth and conclude that it is back in the parent. This is bad.

This is not the nicest fix in the world and it may have small negative impact on performance but I am preoccupied by patching 0.9.x as soon as possible.

A better solution might be to ensure that when a key is not found, we are never right before a key. I had implemented that (it is not hard), but it is more invasive. This patch here is very small.

Fixes  https://github.com/simdjson/simdjson/issues/1521