@lemire 2021-03-31
It causes a significant performance drop (5%) so I will hold on before merging this in our main branch. However, I will patch 0.9.2 right away because a performance drop by 5% is better than buggy code.



Main branch:
```
$ ./buildmaster/benchmark/bench_ondemand --benchmark_filter=ondemand
2021-03-31 17:28:56
Running ./buildmaster/benchmark/bench_ondemand
Run on (16 X 3194.09 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 16384 KiB (x8)
Load Average: 0.31, 0.22, 0.12
simdjson::dom implementation:      haswell
simdjson::ondemand implementation (stage 1): haswell
simdjson::ondemand implementation (stage 2): fallback
----------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------
partial_tweets<simdjson_ondemand>/manual_time             212142 ns       247557 ns         3284 best_branch_miss=1047 best_bytes_per_sec=3.00946G best_cache_miss=0 best_cache_ref=0 best_cycles=712.489k best_cycles_per_byte=1.12822 best_docs_per_sec=4.76547k best_frequency=3.39534G best_instructions=2.41932M best_instructions_per_byte=3.83098 best_instructions_per_cycle=3.39559 best_items_per_sec=476.547k branch_miss=1.10251k bytes=631.515k bytes_per_second=2.77241G/s cache_miss=0 cache_ref=0 cycles=717.984k cycles_per_byte=1.13692 docs_per_sec=4.71382k/s frequency=3.38445G/s instructions=2.41932M instructions_per_byte=3.83098 instructions_per_cycle=3.3696 items=100 items_per_second=471.382k/s [BEST: throughput=  3.01 GB/s doc_throughput=  4765 docs/s instructions=     2419322 cycles=      712489 branch_miss=    1047 cache_miss=       0 cache_ref=         0 items=       100 avg_time=    212142 ns]
Creating a source file spanning 44921 KB
large_random<simdjson_ondemand>/manual_time             57830599 ns     65080338 ns           12 best_branch_miss=845.639k best_bytes_per_sec=815.287M best_cache_miss=0 best_cache_ref=0 best_cycles=191.01M best_cycles_per_byte=4.1525 best_docs_per_sec=17.7241 best_frequency=3.38547G best_instructions=598.053M best_instructions_per_byte=13.0015 best_instructions_per_cycle=3.13101 best_items_per_sec=17.7241M branch_miss=846.4k bytes=45.9988M bytes_per_second=758.558M/s cache_miss=0 cache_ref=0 cycles=195.659M cycles_per_byte=4.25358 docs_per_sec=17.2919/s frequency=3.38332G/s instructions=598.053M instructions_per_byte=13.0015 instructions_per_cycle=3.0566 items=1000k items_per_second=17.2919M/s [BEST: throughput=  0.82 GB/s doc_throughput=    17 docs/s instructions=   598052841 cycles=   191009727 branch_miss=  845639 cache_miss=       0 cache_ref=         0 items=   1000000 avg_time=  57830599 ns]
large_random<simdjson_ondemand_unordered>/manual_time   60427568 ns     67691316 ns           12 best_branch_miss=849.45k best_bytes_per_sec=776.622M best_cache_miss=0 best_cache_ref=0 best_cycles=200.537M best_cycles_per_byte=4.35962 best_docs_per_sec=16.8835 best_frequency=3.38578G best_instructions=613.053M best_instructions_per_byte=13.3276 best_instructions_per_cycle=3.05705 best_items_per_sec=16.8835M branch_miss=849.428k bytes=45.9988M bytes_per_second=725.958M/s cache_miss=0 cache_ref=0 cycles=204.521M cycles_per_byte=4.44622 docs_per_sec=16.5487/s frequency=3.38456G/s instructions=613.053M instructions_per_byte=13.3276 instructions_per_cycle=2.99751 items=1000k items_per_second=16.5487M/s [BEST: throughput=  0.78 GB/s doc_throughput=    16 docs/s instructions=   613052846 cycles=   200537235 branch_miss=  849450 cache_miss=       0 cache_ref=         0 items=   1000000 avg_time=  60427567 ns]
Creating a source file spanning 134087 KB
kostya<simdjson_ondemand>/manual_time                   58537187 ns     79977491 ns           12 best_branch_miss=429.268k best_bytes_per_sec=2.39187G best_cache_miss=0 best_cache_ref=0 best_cycles=194.317M best_cycles_per_byte=1.41522 best_docs_per_sec=17.4201 best_frequency=3.38502G best_instructions=631.028M best_instructions_per_byte=4.59581 best_instructions_per_cycle=3.24741 best_items_per_sec=9.13315M branch_miss=429.117k bytes=137.305M bytes_per_second=2.18451G/s cache_miss=0 cache_ref=0 cycles=198.064M cycles_per_byte=1.44251 docs_per_sec=17.0832/s frequency=3.38356G/s instructions=631.028M instructions_per_byte=4.59581 instructions_per_cycle=3.18598 items=524.288k items_per_second=8.95649M/s [BEST: throughput=  2.39 GB/s doc_throughput=    17 docs/s instructions=   631028288 cycles=   194317133 branch_miss=  429268 cache_miss=       0 cache_ref=         0 items=    524288 avg_time=  58537186 ns]
distinct_user_id<simdjson_ondemand>/manual_time           181382 ns       219389 ns         3857 best_branch_miss=654 best_bytes_per_sec=3.51668G best_cache_miss=0 best_cache_ref=0 best_cycles=609.88k best_cycles_per_byte=0.965741 best_docs_per_sec=5.56864k best_frequency=3.3962G best_instructions=2.00146M best_instructions_per_byte=3.1693 best_instructions_per_cycle=3.28173 best_items_per_sec=640.394k branch_miss=697.886 bytes=631.515k bytes_per_second=3.24256G/s cache_miss=0 cache_ref=0 cycles=613.956k cycles_per_byte=0.972196 docs_per_sec=5.51321k/s frequency=3.38487G/s instructions=2.00146M instructions_per_byte=3.1693 instructions_per_cycle=3.25994 items=115 items_per_second=634.02k/s [BEST: throughput=  3.52 GB/s doc_throughput=  5568 docs/s instructions=     2001463 cycles=      609880 branch_miss=     654 cache_miss=       0 cache_ref=         0 items=       115 avg_time=    181382 ns]
find_tweet<simdjson_ondemand>/manual_time                 132342 ns       169027 ns         5284 best_branch_miss=336 best_bytes_per_sec=4.8051G best_cache_miss=0 best_cache_ref=0 best_cycles=446.394k best_cycles_per_byte=0.706862 best_docs_per_sec=7.60884k best_frequency=3.39654G best_instructions=1.47803M best_instructions_per_byte=2.34046 best_instructions_per_cycle=3.31105 best_items_per_sec=7.60884k branch_miss=357.143 bytes=631.515k bytes_per_second=4.44414G/s cache_miss=0 cache_ref=0 cycles=448.175k cycles_per_byte=0.709682 docs_per_sec=7.55621k/s frequency=3.3865G/s instructions=1.47803M instructions_per_byte=2.34046 instructions_per_cycle=3.29789 items=1 items_per_second=7.55621k/s [BEST: throughput=  4.81 GB/s doc_throughput=  7608 docs/s instructions=     1478034 cycles=      446394 branch_miss=     336 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    132341 ns]
top_tweet<simdjson_ondemand>/manual_time                  182722 ns       219681 ns         3818 best_branch_miss=751 best_bytes_per_sec=3.50223G best_cache_miss=0 best_cache_ref=0 best_cycles=612.334k best_cycles_per_byte=0.969627 best_docs_per_sec=5.54576k best_frequency=3.39586G best_instructions=2.08599M best_instructions_per_byte=3.30316 best_instructions_per_cycle=3.40663 best_items_per_sec=5.54576k branch_miss=773.444 bytes=631.515k bytes_per_second=3.21879G/s cache_miss=0 cache_ref=0 cycles=618.563k cycles_per_byte=0.979491 docs_per_sec=5.4728k/s frequency=3.38527G/s instructions=2.08599M instructions_per_byte=3.30316 instructions_per_cycle=3.37232 items=1 items_per_second=5.4728k/s [BEST: throughput=  3.50 GB/s doc_throughput=  5545 docs/s instructions=     2085993 cycles=      612334 branch_miss=     751 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    182721 ns]
top_tweet<simdjson_ondemand_forward_only>/manual_time     198553 ns       236058 ns         3538 best_branch_miss=804 best_bytes_per_sec=3.22272G best_cache_miss=0 best_cache_ref=0 best_cycles=665.367k best_cycles_per_byte=1.0536 best_docs_per_sec=5.10316k best_frequency=3.39547G best_instructions=2.29458M best_instructions_per_byte=3.63346 best_instructions_per_cycle=3.4486 best_items_per_sec=5.10316k branch_miss=889.711 bytes=631.515k bytes_per_second=2.96215G/s cache_miss=0 cache_ref=0 cycles=672.024k cycles_per_byte=1.06415 docs_per_sec=5.03644k/s frequency=3.38461G/s instructions=2.29458M instructions_per_byte=3.63346 instructions_per_cycle=3.41443 items=1 items_per_second=5.03644k/s [BEST: throughput=  3.22 GB/s doc_throughput=  5103 docs/s instructions=     2294583 cycles=      665367 branch_miss=     804 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    198553 ns]
```

This PR:
```
$ ./buildpr/benchmark/bench_ondemand --benchmark_filter=ondemand
2021-03-31 17:29:15
Running ./buildpr/benchmark/bench_ondemand
Run on (16 X 3194.09 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 16384 KiB (x8)
Load Average: 0.42, 0.25, 0.14
simdjson::dom implementation:      haswell
simdjson::ondemand implementation (stage 1): haswell
simdjson::ondemand implementation (stage 2): fallback
----------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------
partial_tweets<simdjson_ondemand>/manual_time             212611 ns       248099 ns         3286 best_branch_miss=1.21k best_bytes_per_sec=3.01262G best_cache_miss=0 best_cache_ref=0 best_cycles=711.787k best_cycles_per_byte=1.12711 best_docs_per_sec=4.77047k best_frequency=3.39556G best_instructions=2.42278M best_instructions_per_byte=3.83645 best_instructions_per_cycle=3.40379 best_items_per_sec=477.047k branch_miss=1.35865k bytes=631.515k bytes_per_second=2.76629G/s cache_miss=0 cache_ref=0 cycles=716.999k cycles_per_byte=1.13536 docs_per_sec=4.70342k/s frequency=3.37235G/s instructions=2.42278M instructions_per_byte=3.83645 instructions_per_cycle=3.37905 items=100 items_per_second=470.342k/s [BEST: throughput=  3.01 GB/s doc_throughput=  4770 docs/s instructions=     2422775 cycles=      711787 branch_miss=    1210 cache_miss=       0 cache_ref=         0 items=       100 avg_time=    212611 ns]
Creating a source file spanning 44921 KB
large_random<simdjson_ondemand>/manual_time             59914078 ns     66955701 ns           12 best_branch_miss=849.436k best_bytes_per_sec=772.435M best_cache_miss=0 best_cache_ref=0 best_cycles=201.545M best_cycles_per_byte=4.38153 best_docs_per_sec=16.7925 best_frequency=3.38445G best_instructions=599.053M best_instructions_per_byte=13.0232 best_instructions_per_cycle=2.9723 best_items_per_sec=16.7925M branch_miss=849.104k bytes=45.9988M bytes_per_second=732.179M/s cache_miss=0 cache_ref=0 cycles=201.763M cycles_per_byte=4.38627 docs_per_sec=16.6906/s frequency=3.36754G/s instructions=599.053M instructions_per_byte=13.0232 instructions_per_cycle=2.96909 items=1000k items_per_second=16.6906M/s [BEST: throughput=  0.77 GB/s doc_throughput=    16 docs/s instructions=   599052845 cycles=   201545135 branch_miss=  849436 cache_miss=       0 cache_ref=         0 items=   1000000 avg_time=  59914078 ns]
large_random<simdjson_ondemand_unordered>/manual_time   61488891 ns     68577811 ns           11 best_branch_miss=847.405k best_bytes_per_sec=766.223M best_cache_miss=0 best_cache_ref=0 best_cycles=203.206M best_cycles_per_byte=4.41764 best_docs_per_sec=16.6575 best_frequency=3.38489G best_instructions=614.053M best_instructions_per_byte=13.3493 best_instructions_per_cycle=3.02183 best_items_per_sec=16.6575M branch_miss=849.967k bytes=45.9988M bytes_per_second=713.427M/s cache_miss=0 cache_ref=0 cycles=206.917M cycles_per_byte=4.49832 docs_per_sec=16.2631/s frequency=3.36512G/s instructions=614.053M instructions_per_byte=13.3493 instructions_per_cycle=2.96763 items=1000k items_per_second=16.2631M/s [BEST: throughput=  0.77 GB/s doc_throughput=    16 docs/s instructions=   614052846 cycles=   203205838 branch_miss=  847405 cache_miss=       0 cache_ref=         0 items=   1000000 avg_time=  61488891 ns]
Creating a source file spanning 134087 KB
kostya<simdjson_ondemand>/manual_time                   58208606 ns     79436592 ns           12 best_branch_miss=429.197k best_bytes_per_sec=2.37515G best_cache_miss=0 best_cache_ref=0 best_cycles=195.711M best_cycles_per_byte=1.42538 best_docs_per_sec=17.2984 best_frequency=3.38549G best_instructions=632.601M best_instructions_per_byte=4.60727 best_instructions_per_cycle=3.23232 best_items_per_sec=9.06933M branch_miss=429.126k bytes=137.305M bytes_per_second=2.19684G/s cache_miss=0 cache_ref=0 cycles=196.404M cycles_per_byte=1.43042 docs_per_sec=17.1796/s frequency=3.37415G/s instructions=632.601M instructions_per_byte=4.60727 instructions_per_cycle=3.22091 items=524.288k items_per_second=9.00705M/s [BEST: throughput=  2.38 GB/s doc_throughput=    17 docs/s instructions=   632601082 cycles=   195711237 branch_miss=  429197 cache_miss=       0 cache_ref=         0 items=    524288 avg_time=  58208605 ns]
distinct_user_id<simdjson_ondemand>/manual_time           181147 ns       219412 ns         3886 best_branch_miss=677 best_bytes_per_sec=3.55537G best_cache_miss=0 best_cache_ref=0 best_cycles=603.128k best_cycles_per_byte=0.955049 best_docs_per_sec=5.6299k best_frequency=3.39555G best_instructions=2.00692M best_instructions_per_byte=3.17794 best_instructions_per_cycle=3.32752 best_items_per_sec=647.439k branch_miss=812.576 bytes=631.515k bytes_per_second=3.24677G/s cache_miss=0 cache_ref=0 cycles=609.902k cycles_per_byte=0.965776 docs_per_sec=5.52037k/s frequency=3.36688G/s instructions=2.00692M instructions_per_byte=3.17794 instructions_per_cycle=3.29056 items=115 items_per_second=634.842k/s [BEST: throughput=  3.56 GB/s doc_throughput=  5629 docs/s instructions=     2006919 cycles=      603128 branch_miss=     677 cache_miss=       0 cache_ref=         0 items=       115 avg_time=    181147 ns]
find_tweet<simdjson_ondemand>/manual_time                 132738 ns       169801 ns         5283 best_branch_miss=333 best_bytes_per_sec=4.80543G best_cache_miss=0 best_cache_ref=0 best_cycles=446.314k best_cycles_per_byte=0.706735 best_docs_per_sec=7.60937k best_frequency=3.39617G best_instructions=1.47805M best_instructions_per_byte=2.34049 best_instructions_per_cycle=3.31169 best_items_per_sec=7.60937k branch_miss=340.676 bytes=631.515k bytes_per_second=4.43088G/s cache_miss=0 cache_ref=0 cycles=448.381k cycles_per_byte=0.710008 docs_per_sec=7.53366k/s frequency=3.37795G/s instructions=1.47805M instructions_per_byte=2.34049 instructions_per_cycle=3.29642 items=1 items_per_second=7.53366k/s [BEST: throughput=  4.81 GB/s doc_throughput=  7609 docs/s instructions=     1478053 cycles=      446314 branch_miss=     333 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    132737 ns]
top_tweet<simdjson_ondemand>/manual_time                  191024 ns       228513 ns         3566 best_branch_miss=725 best_bytes_per_sec=3.36052G best_cache_miss=0 best_cache_ref=0 best_cycles=638.139k best_cycles_per_byte=1.01049 best_docs_per_sec=5.32136k best_frequency=3.39577G best_instructions=2.09063M best_instructions_per_byte=3.31049 best_instructions_per_cycle=3.27613 best_items_per_sec=5.32136k branch_miss=736.496 bytes=631.515k bytes_per_second=3.0789G/s cache_miss=0 cache_ref=0 cycles=645.549k cycles_per_byte=1.02222 docs_per_sec=5.23493k/s frequency=3.3794G/s instructions=2.09063M instructions_per_byte=3.31049 instructions_per_cycle=3.23853 items=1 items_per_second=5.23493k/s [BEST: throughput=  3.36 GB/s doc_throughput=  5321 docs/s instructions=     2090626 cycles=      638139 branch_miss=     725 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    191024 ns]
top_tweet<simdjson_ondemand_forward_only>/manual_time     202004 ns       239766 ns         3458 best_branch_miss=833 best_bytes_per_sec=3.16781G best_cache_miss=0 best_cache_ref=0 best_cycles=676.899k best_cycles_per_byte=1.07187 best_docs_per_sec=5.0162k best_frequency=3.39546G best_instructions=2.29868M best_instructions_per_byte=3.63995 best_instructions_per_cycle=3.3959 best_items_per_sec=5.0162k branch_miss=869.935 bytes=631.515k bytes_per_second=2.91154G/s cache_miss=0 cache_ref=0 cycles=682.849k cycles_per_byte=1.08129 docs_per_sec=4.95039k/s frequency=3.38037G/s instructions=2.29868M instructions_per_byte=3.63995 instructions_per_cycle=3.36631 items=1 items_per_second=4.95039k/s [BEST: throughput=  3.17 GB/s doc_throughput=  5016 docs/s instructions=     2298680 cycles=      676899 branch_miss=     833 cache_miss=       0 cache_ref=         0 items=         1 avg_time=    202004 ns]
```

This PR (again):

```
$ ./buildpr/benchmark/bench_ondemand --benchmark_filter=ondemand
2021-03-31 17:31:18
Running ./buildpr/benchmark/bench_ondemand
Run on (16 X 3194.09 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 16384 KiB (x8)
Load Average: 0.09, 0.20, 0.13
simdjson::dom implementation:      haswell
simdjson::ondemand implementation (stage 1): haswell
simdjson::ondemand implementation (stage 2): fallback
----------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------
partial_tweets<simdjson_ondemand>/manual_time             210898 ns       245708 ns         3307 best_branch_miss=1.218k best_bytes_per_sec=3.01708G best_cache_miss=0 best_cache_ref=0 best_cycles=710.721k best_cycles_per_byte=1.12542 best_docs_per_sec=4.77753k best_frequency=3.39549G best_instructions=2.42278M best_instructions_per_byte=3.83645 best_instructions_per_cycle=3.4089 best_items_per_sec=477.753k branch_miss=1.24577k bytes=631.515k bytes_per_second=2.78876G/s cache_miss=0 cache_ref=0 cycles=714.18k cycles_per_byte=1.1309 docs_per_sec=4.74162k/s frequency=3.38637G/s instructions=2.42278M instructions_per_byte=3.83645 instructions_per_cycle=3.39239 items=100 items_per_second=474.162k/s [BEST: throughput=  3.02 GB/s doc_throughput=  4777 docs/s instructions=     2422775 cycles=      710721 branch_miss=    1218 cache_miss=       0 cache_ref=         0 items=       100 avg_time=    210898 ns]
Creating a source file spanning 44921 KB
large_random<simdjson_ondemand>/manual_time             59934570 ns     67328870 ns           12 best_branch_miss=849.483k best_bytes_per_sec=792.496M best_cache_miss=0 best_cache_ref=0 best_cycles=196.438M best_cycles_per_byte=4.2705 best_docs_per_sec=17.2286 best_frequency=3.38435G best_instructions=599.053M best_instructions_per_byte=13.0232 best_instructions_per_cycle=3.04958 best_items_per_sec=17.2286M branch_miss=849.766k bytes=45.9988M bytes_per_second=731.929M/s cache_miss=0 cache_ref=0 cycles=202.893M cycles_per_byte=4.41082 docs_per_sec=16.6849/s frequency=3.38523G/s instructions=599.053M instructions_per_byte=13.0232 instructions_per_cycle=2.95256 items=1000k items_per_second=16.6849M/s [BEST: throughput=  0.79 GB/s doc_throughput=    17 docs/s instructions=   599052843 cycles=   196437654 branch_miss=  849483 cache_miss=       0 cache_ref=         0 items=   1000000 avg_time=  59934569 ns]
…